taking care of strange initialisation problem cause by botan library...remove context globals from static/global initialisation because of this with simple trick that requires client applications to define a context provider outside of static/global initialisation

+++ core/testcore/time/testntpclient.cpp
@@ -36,7 +36,7 @@ void TestNtpClient::testWithConnection()
     servers.append("2.pool.ntp.org");
     for (int i = 0; i < 1; i++)
     {
-        qint32 offset;
+        qint64 offset;
         i2pp::core::NtpClient::currentOffset(offset, servers);
     }
 }
